:necktie: Move all AskSeer related items into askSeer.tsx
