+++ static/app/components/searchQueryBuilder/askSeer.tsx
@@ -1,11 +1,64 @@
+import {useRef} from 'react';
 import styled from '@emotion/styled';
+import {useOption} from '@react-aria/listbox';
+import type {ComboBoxState} from '@react-stately/combobox';
 
+import { FeatureBadge } from 'sentry/components/core/badge/featureBadge';
+import InteractionStateLayer from 'sentry/components/core/interactionStateLayer';
+import {useSearchQueryBuilder} from 'sentry/components/searchQueryBuilder/context';
+import {IconSeer} from 'sentry/icons';
+import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
+import {trackAnalytics} from 'sentry/utils/analytics';
+import useOrganization from 'sentry/utils/useOrganization';
 
 export const ASK_SEER_ITEM_KEY = 'ask_seer';
 export const ASK_SEER_CONSENT_ITEM_KEY = 'ask_seer_consent';
 
-export const AskSeerPane = styled('div')`
+function AskSeerOption<T>({state}: {state: ComboBoxState<T>}) {
+  const ref = useRef<HTMLDivElement>(null);
+  const {setDisplaySeerResults} = useSearchQueryBuilder();
+  const organization = useOrganization();
+
+  const {optionProps, labelProps, isFocused, isPressed} = useOption(
+    {
+      key: ASK_SEER_ITEM_KEY,
+      'aria-label': 'Ask Seer',
+      shouldFocusOnHover: true,
+      shouldSelectOnPressUp: true,
+    },
+    state,
+    ref
+  );
+
+  const handleClick = () => {
+    trackAnalytics('trace.explorer.ai_query_interface', {
+      organization,
+      action: 'opened',
+    });
+    setDisplaySeerResults(true);
+  };
+
+  return (
+    <AskSeerListItem ref={ref} onClick={handleClick} {...optionProps}>
+      <InteractionStateLayer isHovered={isFocused} isPressed={isPressed} />
+      <IconSeer />
+      <AskSeerLabel {...labelProps}>
+        {t('Ask Seer')} <FeatureBadge type="beta" />
+      </AskSeerLabel>
+    </AskSeerListItem>
+  );
+}
+
+export function AskSeer<T>({state}: {state: ComboBoxState<T>}) {
+  return (
+    <AskSeerPane>
+      <AskSeerOption state={state} />
+    </AskSeerPane>
+  );
+}
+
+const AskSeerPane = styled('div')`
   grid-area: seer;
   display: flex;
   align-items: center;
@@ -16,7 +69,7 @@ export const AskSeerPane = styled('div')`
   width: 100%;
 `;
 
-export const AskSeerListItem = styled('div')`
+const AskSeerListItem = styled('div')`
   position: relative;
   display: flex;
   align-items: center;
@@ -46,7 +99,7 @@ export const AskSeerListItem = styled('div')`
   }
 `;
 
-export const AskSeerLabel = styled('span')`
+const AskSeerLabel = styled('span')`
   ${p => p.theme.overflowEllipsis};
   color: ${p => p.theme.purple400};
   font-size: ${p => p.theme.fontSize.md};
