+++ static/app/components/searchQueryBuilder/tokens/combobox.tsx
@@ -10,14 +10,13 @@ import {
 import {usePopper} from 'react-popper';
 import styled from '@emotion/styled';
 import {type AriaComboBoxProps} from '@react-aria/combobox';
-import {type AriaListBoxOptions, useOption} from '@react-aria/listbox';
+import {type AriaListBoxOptions} from '@react-aria/listbox';
 import {ariaHideOutside} from '@react-aria/overlays';
 import {mergeRefs} from '@react-aria/utils';
 import {type ComboBoxState, useComboBoxState} from '@react-stately/combobox';
 import type {CollectionChildren, Key, KeyboardEvent} from '@react-types/shared';
 
 import Feature from 'sentry/components/acl/feature';
-import {FeatureBadge} from 'sentry/components/core/badge/featureBadge';
 import {ListBox} from 'sentry/components/core/compactSelect/listBox';
 import type {
   SelectKey,
@@ -29,14 +28,8 @@ import {
 } from 'sentry/components/core/compactSelect/utils';
 import {Input} from 'sentry/components/core/input';
 import {useAutosizeInput} from 'sentry/components/core/input/useAutosizeInput';
-import InteractionStateLayer from 'sentry/components/core/interactionStateLayer';
 import {Overlay} from 'sentry/components/overlay';
-import {
-  ASK_SEER_ITEM_KEY,
-  AskSeerLabel,
-  AskSeerListItem,
-  AskSeerPane,
-} from 'sentry/components/searchQueryBuilder/askSeer';
+import {ASK_SEER_ITEM_KEY, AskSeer} from 'sentry/components/searchQueryBuilder/askSeer';
 import {useSearchQueryBuilder} from 'sentry/components/searchQueryBuilder/context';
 import {useSearchTokenCombobox} from 'sentry/components/searchQueryBuilder/tokens/useSearchTokenCombobox';
 import {
@@ -44,12 +37,8 @@ import {
   itemIsSection,
 } from 'sentry/components/searchQueryBuilder/tokens/utils';
 import type {Token, TokenResult} from 'sentry/components/searchSyntax/parser';
-import {IconSeer} from 'sentry/icons';
-import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import {defined} from 'sentry/utils';
-import {trackAnalytics} from 'sentry/utils/analytics';
-import useOrganization from 'sentry/utils/useOrganization';
 import useOverlay from 'sentry/utils/useOverlay';
 import usePrevious from 'sentry/utils/usePrevious';
 
@@ -260,41 +249,6 @@ function useUpdateOverlayPositionOnContentChange({
   }, [contentRef, isOpen, updateOverlayPosition]);
 }
 
-function AskSeerOption<T>({state}: {state: ComboBoxState<T>}) {
-  const ref = useRef<HTMLDivElement>(null);
-  const {setDisplaySeerResults} = useSearchQueryBuilder();
-  const organization = useOrganization();
-
-  const {optionProps, labelProps, isFocused, isPressed} = useOption(
-    {
-      key: ASK_SEER_ITEM_KEY,
-      'aria-label': 'Ask Seer',
-      shouldFocusOnHover: true,
-      shouldSelectOnPressUp: true,
-    },
-    state,
-    ref
-  );
-
-  const handleClick = () => {
-    trackAnalytics('trace.explorer.ai_query_interface', {
-      organization,
-      action: 'opened',
-    });
-    setDisplaySeerResults(true);
-  };
-
-  return (
-    <AskSeerListItem ref={ref} onClick={handleClick} {...optionProps}>
-      <InteractionStateLayer isHovered={isFocused} isPressed={isPressed} />
-      <IconSeer />
-      <AskSeerLabel {...labelProps}>
-        {t('Ask Seer')} <FeatureBadge type="beta" />
-      </AskSeerLabel>
-    </AskSeerListItem>
-  );
-}
-
 function OverlayContent<T extends SelectOptionOrSectionWithKey<string>>({
   customMenu,
   filterValue,
@@ -349,9 +303,7 @@ function OverlayContent<T extends SelectOptionOrSectionWithKey<string>>({
         />
         {enableAISearch ? (
           <Feature features="organizations:gen-ai-explore-traces">
-            <AskSeerPane>
-              <AskSeerOption state={state} />
-            </AskSeerPane>
+            <AskSeer state={state} />
           </Feature>
         ) : null}
       </ListBoxOverlay>
