+++ static/app/components/searchQueryBuilder/tokens/filterKeyListBox/index.tsx
@@ -7,7 +7,6 @@ import type {ComboBoxState} from '@react-stately/combobox';
 import type {Key} from '@react-types/shared';
 
 import Feature from 'sentry/components/acl/feature';
-import {FeatureBadge} from 'sentry/components/core/badge/featureBadge';
 import {Button} from 'sentry/components/core/button';
 import {ListBox} from 'sentry/components/core/compactSelect/listBox';
 import type {
@@ -16,12 +15,7 @@ import type {
 } from 'sentry/components/core/compactSelect/types';
 import InteractionStateLayer from 'sentry/components/core/interactionStateLayer';
 import {Overlay} from 'sentry/components/overlay';
-import {
-  ASK_SEER_ITEM_KEY,
-  AskSeerLabel,
-  AskSeerListItem,
-  AskSeerPane,
-} from 'sentry/components/searchQueryBuilder/askSeer';
+import {ASK_SEER_ITEM_KEY, AskSeer} from 'sentry/components/searchQueryBuilder/askSeer';
 import {useSearchQueryBuilder} from 'sentry/components/searchQueryBuilder/context';
 import type {CustomComboboxMenuProps} from 'sentry/components/searchQueryBuilder/tokens/combobox';
 import {KeyDescription} from 'sentry/components/searchQueryBuilder/tokens/filterKeyListBox/keyDescription';
@@ -32,12 +26,10 @@ import {
 } from 'sentry/components/searchQueryBuilder/tokens/filterKeyListBox/utils';
 import type {Token, TokenResult} from 'sentry/components/searchSyntax/parser';
 import {getKeyLabel, getKeyName} from 'sentry/components/searchSyntax/utils';
-import {IconMegaphone, IconSeer} from 'sentry/icons';
+import {IconMegaphone} from 'sentry/icons';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
-import {trackAnalytics} from 'sentry/utils/analytics';
 import {useFeedbackForm} from 'sentry/utils/useFeedbackForm';
-import useOrganization from 'sentry/utils/useOrganization';
 import usePrevious from 'sentry/utils/usePrevious';
 
 interface FilterKeyListBoxProps<T> extends CustomComboboxMenuProps<T> {
@@ -144,42 +136,6 @@ function RecentSearchFilterOption<T>({
   );
 }
 
-function AskSeerOption<T>({state}: {state: ComboBoxState<T>}) {
-  const ref = useRef<HTMLDivElement>(null);
-  const {setDisplaySeerResults} = useSearchQueryBuilder();
-  const organization = useOrganization();
-
-  const {optionProps, labelProps, isFocused, isPressed} = useOption(
-    {
-      key: ASK_SEER_ITEM_KEY,
-      'aria-label': 'Ask Seer',
-      shouldFocusOnHover: true,
-      shouldSelectOnPressUp: true,
-    },
-    state,
-    ref
-  );
-
-  const handleClick = () => {
-    trackAnalytics('trace.explorer.ai_query_interface', {
-      organization,
-      action: 'opened',
-    });
-    setDisplaySeerResults(true);
-  };
-
-  return (
-    <AskSeerListItem ref={ref} onClick={handleClick} {...optionProps}>
-      <InteractionStateLayer isHovered={isFocused} isPressed={isPressed} />
-      <IconSeer />
-      <AskSeerLabel {...labelProps}>
-        {t('Ask Seer')}
-        <FeatureBadge type="beta" />
-      </AskSeerLabel>
-    </AskSeerListItem>
-  );
-}
-
 function useHighlightFirstOptionOnSectionChange({
   state,
   selectedSection,
@@ -271,9 +227,7 @@ function FilterKeyMenuContent<T extends SelectOptionOrSectionWithKey<string>>({
     <Fragment>
       {enableAISearch ? (
         <Feature features="organizations:gen-ai-explore-traces">
-          <AskSeerPane>
-            <AskSeerOption state={state} />
-          </AskSeerPane>
+          <AskSeer state={state} />
         </Feature>
       ) : null}
       {showRecentFilters ? (
